Generator for discontinuous datestamped data
This generator is used when you want to plot non-continuous datestamped data. For
example, you might have data spanning three weeks, but you're missing data for the
weekends, and you want to express that those data points are missing. If a queryset
doesn't have a value for a particular date, then this generator yields a 'null'.